Move Cachable and LockableQuery Object interfaces to own file
