+++ rotkehlchen/exchanges/exchange.py
@@ -9,7 +9,7 @@ from rotkehlchen.exchanges.data_structures import AssetMovement, MarginPosition,
 from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.serialization.deserialize import deserialize_location
 from rotkehlchen.typing import ApiKey, ApiSecret, T_ApiKey, T_ApiSecret, Timestamp
-from rotkehlchen.utils.misc import CacheableObject, LockableQueryObject
+from rotkehlchen.utils.interfaces import CacheableObject, LockableQueryObject
 
 if TYPE_CHECKING:
     from rotkehlchen.db.dbhandler import DBHandler
