+++ rotkehlchen/exchanges/poloniex.py
@@ -54,12 +54,8 @@ from rotkehlchen.typing import (
     TradePair,
 )
 from rotkehlchen.user_messages import MessagesAggregator
-from rotkehlchen.utils.misc import (
-    cache_response_timewise,
-    create_timestamp,
-    protect_with_lock,
-    retry_calls,
-)
+from rotkehlchen.utils.interfaces import cache_response_timewise, protect_with_lock
+from rotkehlchen.utils.misc import create_timestamp, retry_calls
 from rotkehlchen.utils.serialization import rlk_jsonloads, rlk_jsonloads_dict, rlk_jsonloads_list
 
 if TYPE_CHECKING:
